* main.c: But the GNULIB gethostname accepts an int not ssize_t.
CVS-ID: src/ChangeLog 1.2508
CVS-ID: src/main.c 1.189

+++ src/main.c
@@ -18,7 +18,7 @@
 #ifdef HAVE_WINSOCK_H
 # include <winsock.h>
 #elif !HAVE_GETHOSTNAME
-extern int gethostname (char *, size_t);
+extern int gethostname (char *, int);
 #endif
 
 char *program_name;
